Broken implementation of splitting slots n signals
