+++ cockpit/IPC/src/views/IODebug.tsx
@@ -2,39 +2,24 @@
 /* eslint-disable react/jsx-no-undef */
 /* eslint-disable react/no-unstable-nested-components */
 import React, {
-  ReactElement, useEffect, useRef, useState,
+  useEffect, useRef, useState,
 } from 'react';
 import {
   DataList,
-  DataListItem,
-  DataListItemCells,
-  DataListItemRow,
-  DataListCell,
-  DataListAction,
-  Dropdown,
-  DropdownItem,
   Title,
-  Tooltip,
   DrawerContent,
   Drawer,
   DrawerContentBody,
   DrawerPanelContent,
-  MenuToggleElement,
-  DropdownList,
   Spinner,
 } from '@patternfly/react-core';
 import { loadExternalScript } from 'src/Components/Interface/ScriptLoader';
 import './IODebug.css';
-import Circle from 'src/Components/Simple/Circle';
 import DynamicNavbar from 'src/Components/NavBar/DynamicNavBar';
+import ListItem from 'src/Components/IOList/ListItem';
 import Hamburger from 'hamburger-react';
-import CustomMenuToggle from 'src/Components/Dropdown/CustomMenuToggle';
 import { DarkModeType } from 'src/App';
-import StringTinker from 'src/Components/Tinker/StringTinker';
-import BoolTinker from 'src/Components/Tinker/BoolTinker';
-import { removeSlotOrg } from 'src/Components/Form/WidgetFunctions';
 import { TFC_DBUS_DOMAIN, TFC_DBUS_ORGANIZATION } from 'src/variables';
-import NumberTinker from 'src/Components/Tinker/NumberTinker';
 
 declare global {
   interface Window { cockpit: any; }
@@ -63,15 +48,16 @@ const parseXMLInterfaces = (xml: string): { name: string, valueType: string }[] 
 };
 
 // eslint-disable-next-line react/function-component-definition
-const IODebug:React.FC<DarkModeType> = ({ isDark }) => {
+const IODebug: React.FC<DarkModeType> = ({ isDark }) => {
   const [dbusInterfaces, setDbusInterfaces] = useState<any[]>([]);
   const [processes, setProcesses] = useState<string[]>();
   const [isDrawerExpanded, setIsDrawerExpanded] = useState<boolean>(true);
   const [activeDropdown, setActiveDropdown] = useState<number | null>(null);
   // eslint-disable-next-line @typescript-eslint/comma-spacing
-  const eventHandlersRef = useRef<Map<string,(e:any) => void >>(new Map());
+  const eventHandlersRef = useRef<Map<string,(e: any) => void>>(new Map()); // NOSONAR
 
-  const path = `/${TFC_DBUS_DOMAIN}/${TFC_DBUS_ORGANIZATION}/Slots`;
+  const slotPath = `/${TFC_DBUS_DOMAIN}/${TFC_DBUS_ORGANIZATION}/Slots`;
+  const signalPath = `/${TFC_DBUS_DOMAIN}/${TFC_DBUS_ORGANIZATION}/Signals`;
 
   useEffect(() => {
     if (!processes) return;
@@ -93,14 +79,14 @@ const IODebug:React.FC<DarkModeType> = ({ isDark }) => {
       // eslint-disable-next-line no-restricted-syntax
       for (const process of processes) {
         const processDBUS = window.cockpit.dbus(process, { bus: 'system', superuser: 'try' });
-        const processProxy = processDBUS.proxy('org.freedesktop.DBus.Introspectable', path);
+        const slotProcessProxy = processDBUS.proxy('org.freedesktop.DBus.Introspectable', slotPath);
         try {
           // eslint-disable-next-line no-await-in-loop, @typescript-eslint/no-loop-func
-          await processProxy.call('Introspect').then((data:any) => {
+          await slotProcessProxy.call('Introspect').then((data: any) => {
             const interfacesData = parseXMLInterfaces(data);
             // eslint-disable-next-line no-restricted-syntax
             for (const interfaceData of interfacesData) {
-              const proxy = processDBUS.proxy(interfaceData.name, path);
+              const proxy = processDBUS.proxy(interfaceData.name, slotPath);
               proxy.wait().then(() => {
                 const handler = handleChanged(interfaceData.name);
                 proxy.addEventListener('changed', handler);
@@ -110,7 +96,34 @@ const IODebug:React.FC<DarkModeType> = ({ isDark }) => {
                   process,
                   interfaceName: interfaceData.name,
                   type: interfaceData.valueType,
-                  forcestate: null,
+                  direction: 'slot',
+                  hidden: false,
+                });
+              });
+            }
+          });
+        } catch (e) {
+          console.log(e);
+        }
+
+        const signalProcessProxy = processDBUS.proxy('org.freedesktop.DBus.Introspectable', signalPath);
+        try {
+          // eslint-disable-next-line no-await-in-loop, @typescript-eslint/no-loop-func
+          await signalProcessProxy.call('Introspect').then((data: any) => {
+            const interfacesData = parseXMLInterfaces(data);
+            // eslint-disable-next-line no-restricted-syntax
+            for (const interfaceData of interfacesData) {
+              const proxy = processDBUS.proxy(interfaceData.name, signalPath);
+              proxy.wait().then(() => {
+                const handler = handleChanged(interfaceData.name);
+                proxy.addEventListener('changed', handler);
+                eventHandlersRef.current.set(interfaceData.name, handler);
+                interfaces.push({
+                  proxy,
+                  process,
+                  interfaceName: interfaceData.name,
+                  type: interfaceData.valueType,
+                  direction: 'signal',
                   hidden: false,
                 });
               });
@@ -129,7 +142,7 @@ const IODebug:React.FC<DarkModeType> = ({ isDark }) => {
   useEffect(() => {
     const callback = (allNames: string[]) => {
       setProcesses(
-        allNames.filter((name:string) => name.includes(`${TFC_DBUS_DOMAIN}.${TFC_DBUS_ORGANIZATION}.tfc.`)),
+        allNames.filter((name: string) => name.includes(`${TFC_DBUS_DOMAIN}.${TFC_DBUS_ORGANIZATION}.tfc.`)),
       );
     };
     loadExternalScript(callback);
@@ -148,127 +161,15 @@ const IODebug:React.FC<DarkModeType> = ({ isDark }) => {
   };
 
   /**
-   * Handles the content of the secondary column for booleans
-   * @param data The data to be displayed
-   * @returns ReactElement to be displayed
-   */
-  function handleBoolContent(data: any): ReactElement {
-    return (
-      <Tooltip
-        content={`Value is ${data.Value ? 'true' : 'false'}`}
-        enableFlip
-        distance={5}
-        entryDelay={1000}
-      >
-        <Circle size="1rem" color={data.Value ? 'green' : 'red'} />
-      </Tooltip>
-    );
-  }
-
-  /**
-   * Handles the content of the secondary column for strings
-   * @param data The data to be displayed
-   * @returns ReactElement to be displayed
-   */
-  function handleStringContent(data: any): ReactElement {
-    return (
-      <p style={{ marginBottom: '0px' }}>{data.Value}</p>
-    );
-  }
-
-  /**
    * Toggles the visibility of IO for the given process
    * @param selected The selected process
    */
-  function toggleSelection(selected:string) {
+  function toggleSelection(selected: string) {
     const updatedData = dbusInterfaces.map((dbusInterface) => ({
       ...dbusInterface,
       hidden: selected && dbusInterface.process !== selected,
     }));
     setDbusInterfaces(updatedData);
-  }
-
-  /**
-   * Handles the content of the secondary column
-   * @param data Interface data
-   * @returns ReactElement to be displayed
-   */
-  function getSecondaryContent(data: any): ReactElement | null {
-    const internals = (interfacedata: any) => {
-      switch (interfacedata.type) {
-        case 'b':
-          return handleBoolContent(interfacedata.proxy.data);
-
-        case 's':
-          return handleStringContent(interfacedata.proxy.data);
-
-        case 'n': // INT16
-        case 'q': // UINT16
-        case 'i': // INT32
-        case 'u': // UINT32
-        case 'x': // INT64
-        case 't': // UNIT64
-        case 'd': // Double
-        case 'y': // Byte
-          return handleStringContent(interfacedata.proxy.data);
-
-        default:
-          return <>Type Error</>;
-      }
-    };
-
-    return (
-      <div style={{
-        height: '100%',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'end',
-      }}
-      >
-        {internals(data)}
-      </div>
-    );
-  }
-
-  /**
-   * Handles the content of the third column to enable tinkering
-   * @param data Interface data
-   * @returns ReactElement to be displayed
-   */
-  function getTinkerInterface(data:any): React.ReactElement | null {
-    const internals = (interfacedata: any) => {
-      switch (interfacedata.type) {
-        case 'b':
-          return <BoolTinker data={interfacedata} />;
-
-        case 's':
-          return <StringTinker data={interfacedata} />;
-
-        case 'n': // INT16
-        case 'q': // UINT16
-        case 'i': // INT32
-        case 'u': // UINT32
-        case 'x': // INT64
-        case 't': // UNIT64
-        case 'd': // Double
-        case 'y': // Byte
-          return <NumberTinker data={interfacedata} />;
-
-        default:
-          return <>Type Error</>;
-      }
-    };
-    return (
-      <div style={{
-        height: '100%',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'start',
-      }}
-      >
-        {internals(data)}
-      </div>
-    );
   }
 
   /**
@@ -343,70 +244,14 @@ const IODebug:React.FC<DarkModeType> = ({ isDark }) => {
                 <DataList aria-label="Checkbox and action data list example">
                   {dbusInterfaces.length > 0 ? dbusInterfaces.map((dbusInterface: any, index: number) => {
                     if (!dbusInterface.hidden) {
-                      return (
-                        <DataListItem aria-labelledby="check-action-item1" key={dbusInterface.proxy.iface + dbusInterface.process}>
-                          <DataListItemRow size={10}>
-                            <DataListItemCells
-                              dataListCells={[
-                                <DataListCell key="primary content" style={{ textAlign: 'left', maxWidth: '30rem' }}>
-                                  <p style={{ minWidth: '30rem' }}>{removeSlotOrg(dbusInterface.proxy.iface)}</p>
-                                </DataListCell>,
-                                <DataListCell
-                                  key="secondary content 1"
-                                  style={{
-                                    textAlign: 'right',
-                                    height: '100%',
-                                    display: 'flex',
-                                    alignItems: 'center',
-                                  }}
-                                >
-                                  {getSecondaryContent(dbusInterface)}
-                                </DataListCell>,
-                                <DataListCell
-                                  key="secondary content 2"
-                                  style={{
-                                    textAlign: 'right',
-                                    height: '100%',
-                                    display: 'flex',
-                                    alignItems: 'center',
-                                  }}
-                                >
-                                  {getTinkerInterface(dbusInterface)}
-                                </DataListCell>,
-                              ]}
-                            />
-                            <DataListAction
-                              aria-labelledby="check-action-item1 check-action-action1"
-                              id="check-action-action1"
-                              aria-label="Actions"
-                              isPlainButtonAction
-                            >
-                              <Dropdown
-                                className="DropdownItem"
-                                key={`${dbusInterface.proxy.iface}${dbusInterface.process}`}
-                                toggle={(toggleRef: React.Ref<MenuToggleElement>) => ( // NOSONAR
-                                  <CustomMenuToggle
-                                    toggleRef={(ref) => {
-                                      if (typeof toggleRef === 'function') {
-                                        toggleRef(ref);
-                                      }
-                                      dropdownRefs.current[index] = ref; // Store the ref for the dropdown
-                                    }}
-                                    onClick={() => onToggleClick(index)}
-                                    isExpanded={dbusInterface.dropdown}
-                                  />
-                                )}
-                                isOpen={activeDropdown === index}
-                              >
-                                <DropdownList>
-                                  <DropdownItem key="tinker" style={{ textDecoration: 'none' }}> Tinker </DropdownItem>
-                                  <DropdownItem key="history" style={{ textDecoration: 'none' }} isDisabled> View History </DropdownItem>
-                                </DropdownList>
-                              </Dropdown>
-                            </DataListAction>
-                          </DataListItemRow>
-                        </DataListItem>
-                      );
+                      <ListItem
+                        dbusInterface={dbusInterface}
+                        index={index}
+                        direction={dbusInterface.direction}
+                        activeDropdown={activeDropdown}
+                        dropdownRefs={dropdownRefs}
+                        onToggleClick={onToggleClick}
+                      />;
                     }
                     return null;
                   })
